Refactor GameCreatorPage into wizard stages and implement game flow stage.
